Refactor: 컴포넌트 파일 리팩토링 (#152)
