+++ src/features/review/ui/ReviewThird.jsx
@@ -1,8 +1,8 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import styled from 'styled-components'
 import RangeItem from './RangeItem.jsx';
 import { ToggleCheckbox } from "./ToggleCheckBox.jsx";
-import { GuideMsg, ImgSection, MsgWrapper, Scroll, ThemeImg, ThemeSubText, ThemeTitle, Wrap1, Wrap2, Wrap3 } from "../../../shared/components/ReviewStyle.js";
+import { GuideMsg, ImgSection, MsgWrapper, Scroll, ThemeImg, ThemeSubText, ThemeTitle, Wrap1, Wrap2, Wrap3 } from "../../../shared/styles/ReviewStyles.js";
 import InputBox from "./InputBox.jsx";
 import useDevice from "../../../shared/hooks/useDevice.js";
 import { useRecoilState } from 'recoil';
@@ -11,7 +11,7 @@ import { reviewState } from '../../themeDetail/model/reviewAtom.jsx';
 export default function ReviewThird({themeData}) {
 
   // 반응형
-  const { isDesktop, isTablet, isMobile } = useDevice();
+  const { isMobile } = useDevice();
   
   // 후기 데이터 상태
   const [review, setReview] = useRecoilState(reviewState);
@@ -350,17 +350,4 @@ const Wrap = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
-`;
-
-const ItemText2 = styled.div`
-  color: var(--RIU_Primary-100, #718FF2);
-  text-align: center;
-  font-family: Pretendard-Bold;
-  font-size: 0.75em;
-
-  @media (max-width: 1024px) {
-  }
-  @media (max-width: 768px) {
-    font-size: 0.625em;
-  }
 `;
