+++ src/shared/components/SearchInput.jsx
@@ -1,13 +1,14 @@
 import PropTypes from "prop-types";
 import styled from 'styled-components';
 import { useState } from 'react';
-import SearchIcon from '../../shared/assets/icons/common/searchIcon.svg?react';
-import CancelIcon from '../../shared/assets/icons/common/cancelIcon.svg?react';
+import SearchIcon from '../assets/icons/common/searchIcon.svg?react';
+import CancelIcon from '../assets/icons/common/cancelIcon.svg?react';
 
 function SearchInput({ type, keyword, setKeyword, onSearch }) {
 
   const [expanded, setExpanded] = useState(type !== 'header');
 
+  // 검색 버튼 클릭
   const handleSearchClick = (e) => {
     e.stopPropagation();
     if (!expanded) {
