+++ src/shared/utils/scrollToTopFunc.js
